+++ src/main/java/dnaaaaahtac/wooriforei/domain/openapi/dto/information/InformationDetailDTO.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
 @Getter
-public class InformationDetailDto {
+public class InformationDetailDTO {
 
     @JsonProperty("TRSMICNM")
     private String trsmicnm;
