+++ src/main/java/dnaaaaahtac/wooriforei/domain/openapi/dto/seoulgoods/SeoulGoodsDetailDTO.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
 @Getter
-public class SeoulGoodsDetailDto {
+public class SeoulGoodsDetailDTO {
 
     @JsonProperty("NM")
     private String nm;
