+++ src/main/java/dnaaaaahtac/wooriforei/domain/openapi/dto/seoulgoods/SeoulGoodsResponseDTO.java
@@ -1,13 +1,12 @@
 package dnaaaaahtac.wooriforei.domain.openapi.dto.seoulgoods;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
 import lombok.Getter;
 
 @Getter
-public class SeoulGoodsResponseDto {
+public class SeoulGoodsResponseDTO {
 
     @JsonProperty("frgnrTourGiftShopInfo")
-    private SeoulGoodsResponseVo frgnrtourgiftshopinfo;
+    private SeoulGoodsResponseVO frgnrtourgiftshopinfo;
 
 }
